Parse onion addresses and ToString
